requirements.txt
fastapi==0.116.1
firebase_admin==7.1.0
pydantic==2.11.7
python-dotenv==1.1.1
uvicorn==0.35.0
validate_docbr==1.11.1
